### 2018-04-06 #666 @whedon
[PRE REVIEW]: SuchTree : A Python library for fast, thread-safe computations on phylogenetic trees
**Submitting author:** @ryneches (<a href="http://orcid.org/0000-0002-2055-8381">Russell Y. Neches</a>)
**Repository:** <a href="https://github.com/ryneches/SuchTree" target ="_blank">https://github.com/ryneches/SuchTree</a>
**Version:** 0.6
