### 2018-04-10 366ff76
found another type testing bug
