+++ SuchTree/__init__.py
@@ -1,3 +1,3 @@
 #!/usr/bin/env python
 from __future__ import absolute_import
-from SuchTree.SuchTree import SuchTree, SuchLinkedTrees, pearson
+from SuchTree.SuchTree import SuchTree, SuchLinkedTrees
