### 2018-04-15 #666 @ryneches
Hey @karthik! Here are some suggested reviewers :
* @giraola (from the JOSS list of reviewers)
* @arvestad (or perhaps someone in his lab?)
* @trayc7 (or someone in her lab?)
* @snacktavish (or someone in her lab?)
