### 2018-04-15 #666 @arvestad
Never been included in reviewer discussions like this before. :-)

	L

> On 15 Apr 2018, at 07:35, Russell Neches <notifications@github.com> wrote:
>
> Hey @karthik <https://github.com/karthik>! Here are some suggested reviewers :
>
> @giraola <https://github.com/giraola> (from the JOSS list of reviewers)
> @arvestad <https://github.com/arvestad> (or perhaps someone in his lab?)
> @trayc7 <https://github.com/trayc7> (or someone in her lab?)
> @snacktavish <https://github.com/snacktavish> (or someone in her lab?)
