### 2018-04-16 #666 @arvestad
I don't mind at all. I am new to JOSS and am curious about its extreme openness.
