### 2018-04-27 #678 @ryneches
Article proof looks good to me.
