### 2018-04-28 #678 @giraola
@ryneches please add DOIs to all references listed in the manuscript.
