### 2018-04-30 8617d28
updated bibliography in JOSS paper
