+++ docs/paper.bib
@@ -6,7 +6,8 @@
   number={12},
   pages={1569--1571},
   year={2010},
-  publisher={Oxford Univ Press}
+  publisher={Oxford Univ Press},
+  doi={https://doi.org/10.1093/bioinformatics/btq228}
 }
 
 @article{ete3,
@@ -17,7 +18,8 @@
   number={6},
   pages={1635--1638},
   year={2016},
-  publisher={Society for Molecular Biology and Evolution}
+  publisher={Society for Molecular Biology and Evolution},
+  doi={https://doi.org/10.1093/molbev/msw046}
 }
 
 @article{pylogeny,
@@ -27,7 +29,8 @@
   volume={1},
   pages={e9},
   year={2015},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi={https://doi.org/10.7717/peerj-cs.9}
 }
 
 @article{biophylo,
@@ -38,7 +41,8 @@
   number={1},
   pages={209},
   year={2012},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={https://doi.org/10.1186/1471-2105-13-209}
 }
 
 @misc{skbio,
@@ -49,4 +53,28 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/biocore/scikit-bio/}},
   commit = {cd2725074ea55838fa357d341c4f214196ed017a}
+}
+
+@article{price2009fasttree,
+  title={FastTree: computing large minimum evolution trees with profiles instead of a distance matrix},
+  author={Price, Morgan N and Dehal, Paramvir S and Arkin, Adam P},
+  journal={Molecular biology and evolution},
+  volume={26},
+  number={7},
+  pages={1641--1650},
+  year={2009},
+  publisher={Oxford University Press},
+  doi={https://doi.org/10.1093/molbev/msp077}
+}
+
+@article{price2010fasttree,
+  title={FastTree 2--approximately maximum-likelihood trees for large alignments},
+  author={Price, Morgan N and Dehal, Paramvir S and Arkin, Adam P},
+  journal={PloS one},
+  volume={5},
+  number={3},
+  pages={e9490},
+  year={2010},
+  publisher={Public Library of Science},
+  doi={https://doi.org/10.1371/journal.pone.0009490}
 }
