### 2018-05-01 49945da
tweak JOSS BibTeX
