+++ docs/paper.bib
@@ -70,7 +70,7 @@
 @article{price2010fasttree,
   title={FastTree 2--approximately maximum-likelihood trees for large alignments},
   author={Price, Morgan N and Dehal, Paramvir S and Arkin, Adam P},
-  journal={PloS one},
+  journal={PLOS ONE},
   volume={5},
   number={3},
   pages={e9490},
