### 2018-05-01 #678 @ryneches
Hmm. I'm not happy with the way the references are getting formatted. Multiple references look weird, and BibTeX is doing that thing where it turns repeated author lists into a long dash. :-/

Is there a formatting guide? I'm not so familiar with the MarkDown to LaTeX way of doing things.
