### 2018-05-03 #678 @arfon
> Hmm. I'm not happy with the way the references are getting formatted. Multiple references look weird, and BibTeX is doing that thing where it turns repeated author lists into a long dash. :-/

This is the standard style for our bibliography - it's because the author list of the second paper is the same as the one above it.
