### 2018-05-18 #678 @giraola
@ryneches please add installation guidelines and list SuchTree dependencies in the readme file.
