### 2018-05-20 4a639b1
added dependencies & install instructions
