### 2018-05-20 5d6ad02
moved install instructions next to usage
