### 2018-06-12 #678 @giraola
@ryneches I just completed the revision. Sorry for this delay.
