### 2018-06-16 902e7e0
updating example notebook, bigtree data
