### 2018-06-19 #678 @karthik
@arfon Yes, definitely. I communicated with Russel before I left for my trip and am just returning to editorial duties. Will finish this up asap.
