### 2018-06-20 c484060
updating example datasets and notebooks
