+++ data/bigtrees/README.md
@@ -4,3 +4,8 @@ Two rather large trees containing the same taxa.
 
 * **ml.tree** : 108653 nodes, 54327 taxa, approximate maximum-likelihood
 * **nj.tree** : 108653 nodes, 54327 taxa, neighbor-joining
+
+A multi-species microbiome.
+* **host.tree** : 206891 nodes, 103446 taxa, approximate maximum-likelihood
+* **guest.tree** : 27 nodes, 14 taxa, RAxML
+* **links.csv** : 14 by 103446 observation matrix
