### 2018-06-22 #678 @arfon
@ryneches - I made a small fix to one of your bibtex entries which wasn't working in https://github.com/ryneches/SuchTree/pull/14
